added style to checkboxes on traits filter
